Pull `sc_consensus_subspace::authorship` module into `sc_consensus_subspace::slot_worker` module
